Fix LP64 in Build_Frame offset array and keyframe palette calc
- Build_Frame offset[]: unsigned long → uint32_t (file has 4-byte
  entries; was reading 56 bytes into 28-byte data causing corruption)
- Mem_Copy sizes: sizeof(unsigned long) → sizeof(uint32_t) for all
  file-format offset reads in keyframe delta chain
- Get_Build_Frame_Palette: sizeof(unsigned long) → sizeof(uint32_t)
  in palette offset calculation
- CC_Draw_Shape: detect SHP format before calling Build_Frame to
  prevent KeyFrame processing of SHP data; add frame dimension check

+++ modern.src/keyframe.cpp
@@ -215,13 +215,13 @@ void Enable_Uncompressed_Shapes (void)
 unsigned long Build_Frame(void const *dataptr, unsigned short framenumber, void *buffptr)
 {
 	char *ptr, *lockptr;//, *uncomp_ptr;
-	unsigned long offset[SUBFRAMEOFFS];
-	unsigned long offcurr, off16, offdiff;
+	uint32_t offset[SUBFRAMEOFFS]; /* LP64: was unsigned long but file has 4-byte entries */
+	uint32_t offcurr, off16, offdiff;
 	KeyFrameHeaderType *keyfr;
 	unsigned short buffsize, currframe, subframe;
-	unsigned long length = 0;
+	unsigned long length = 0; /* LP64: stays unsigned long — used as return value (pointer) */
 	char frameflags;
-	unsigned long return_value;
+	unsigned long return_value; /* LP64: holds pointer cast to unsigned long */
 	char *temp_shape_ptr;
 
 	//
@@ -361,7 +361,7 @@ unsigned long Build_Frame(void const *dataptr, unsigned short framenumber, void 
 			currframe = (unsigned short)offset[1];
 
 			ptr = (char *)Add_Long_To_Pointer( dataptr, (((unsigned long)currframe << 3) + sizeof(KeyFrameHeaderType)) );
-			Mem_Copy( ptr, &offset[0], (long)(SUBFRAMEOFFS * sizeof(unsigned long)) );
+			Mem_Copy( ptr, &offset[0], (long)(SUBFRAMEOFFS * sizeof(uint32_t)) ); /* LP64: file has 4-byte entries */
 		}
 
 		// key frame
@@ -425,7 +425,7 @@ unsigned long Build_Frame(void const *dataptr, unsigned short framenumber, void 
 					Mem_Copy( Add_Long_To_Pointer( dataptr,
 									(((unsigned long)currframe << 3) +
 									sizeof(KeyFrameHeaderType)) ),
-						&offset[0], (long)(SUBFRAMEOFFS * sizeof(unsigned long)) );
+						&offset[0], (long)(SUBFRAMEOFFS * sizeof(uint32_t)) ); /* LP64: 4-byte file entries */
 					subframe = 0;
 				}
 			}
@@ -593,7 +593,7 @@ bool Get_Build_Frame_Palette(void const * dataptr, void * palette)
 {
 	if (dataptr && (((KeyFrameHeaderType const *)dataptr)->flags & 1)) {
 		char const * ptr = (char const *)Add_Long_To_Pointer( dataptr,
-							( (( (long)sizeof(unsigned long) << 1) *
+							( (( (long)sizeof(uint32_t) << 1) * /* LP64: was sizeof(unsigned long), file has 4-byte entries */
 								((KeyFrameHeaderType *) dataptr)->frames ) +
 							16 + sizeof(KeyFrameHeaderType) ) );
